+++ src/main/java/shop/controller/AuthController.java
@@ -6,8 +6,8 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import shop.dto.request.RegisterUserRequestDto;
-import shop.dto.responce.RegisterUserResponseDto;
+import shop.dto.request.UserRegistrationRequestDto;
+import shop.dto.responce.UserResponseDto;
 import shop.exception.RegistrationException;
 import shop.service.UserService;
 
@@ -18,9 +18,9 @@ public class AuthController {
     private final UserService userService;
 
     @PostMapping("/registration")
-    public RegisterUserResponseDto register(
-            @Valid @RequestBody RegisterUserRequestDto registerUserRequestDto)
+    public UserResponseDto register(
+            @Valid @RequestBody UserRegistrationRequestDto userRegistrationRequestDto)
             throws RegistrationException {
-        return userService.register(registerUserRequestDto);
+        return userService.register(userRegistrationRequestDto);
     }
 }
