+++ src/main/java/shop/dto/request/UserRegistrationRequestDto.java
@@ -4,7 +4,7 @@ import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.Setter;
-import shop.annatation.FieldMatch;
+import shop.validation.FieldMatch;
 
 @Setter
 @Getter
@@ -13,7 +13,7 @@ import shop.annatation.FieldMatch;
         secondFieldName = "repeatPassword",
         message = "Password must match!"
 )
-public class RegisterUserRequestDto {
+public class UserRegistrationRequestDto {
     @NotBlank
     @Email
     private String email;
