+++ src/main/java/shop/dto/responce/BookDto.java
@@ -1,8 +1,5 @@
 package shop.dto.responce;
 
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Positive;
 import java.math.BigDecimal;
 import lombok.Getter;
 import lombok.Setter;
@@ -10,16 +7,10 @@ import lombok.Setter;
 @Setter
 @Getter
 public class BookDto {
-    @NotBlank
     private Long id;
-    @NotBlank
     private String title;
-    @NotBlank
     private String author;
-    @NotBlank
     private String isbn;
-    @NotNull
-    @Positive
     private BigDecimal price;
     private String description;
     private String coverImage;
