+++ src/main/java/shop/dto/responce/UserResponseDto.java
@@ -1,21 +1,14 @@
 package shop.dto.responce;
 
-import jakarta.validation.constraints.Email;
-import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.Setter;
 
 @Setter
 @Getter
-public class RegisterUserResponseDto {
-    @NotBlank
+public class UserResponseDto {
     private Long id;
-    @NotBlank
-    @Email
     private String email;
-    @NotBlank
     private String firstName;
-    @NotBlank
     private String lastName;
     private String shippingAddress;
 }
