+++ src/main/java/shop/entity/User.java
@@ -17,23 +17,13 @@ public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    @Column(
-            unique = true,
-            nullable = false
-    )
+    @Column(unique = true, nullable = false)
     private String email;
     @Column(nullable = false)
     private String password;
-    @Column(
-            name = "first_name",
-            nullable = false
-    )
+    @Column(nullable = false)
     private String firstName;
-    @Column(
-            name = "last_name",
-            nullable = false
-    )
+    @Column(nullable = false)
     private String lastName;
-    @Column(name = "shipping_address")
     private String shippingAddress;
 }
