+++ src/main/java/shop/repository/user/UserRepository.java
@@ -1,9 +1,8 @@
 package shop.repository.user;
 
-import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import shop.entity.User;
 
 public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
+    boolean existsUserByEmail(String email);
 }
