+++ src/main/java/shop/service/impl/UserServiceImpl.java
@@ -2,8 +2,8 @@ package shop.service.impl;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import shop.dto.request.RegisterUserRequestDto;
-import shop.dto.responce.RegisterUserResponseDto;
+import shop.dto.request.UserRegistrationRequestDto;
+import shop.dto.responce.UserResponseDto;
 import shop.entity.User;
 import shop.exception.RegistrationException;
 import shop.mapper.UserMapper;
@@ -17,13 +17,13 @@ public class UserServiceImpl implements UserService {
     private final UserMapper userMapper;
 
     @Override
-    public RegisterUserResponseDto register(RegisterUserRequestDto registerUserRequestDto)
+    public UserResponseDto register(UserRegistrationRequestDto registerUserRequestDto)
             throws RegistrationException {
-        if (userRepository.findByEmail(registerUserRequestDto.getEmail()).isPresent()) {
+        if (userRepository.existsUserByEmail(registerUserRequestDto.getEmail())) {
             throw new RegistrationException("Can't register user with given email");
         }
         User user = userMapper.toUser(registerUserRequestDto);
-        User savedUser = userRepository.save(user);
-        return userMapper.toRegisterUserResponseDto(savedUser);
+        userRepository.save(user);
+        return userMapper.toRegisterUserResponseDto(user);
     }
 }
