+++ src/main/java/shop/validation/FieldMatch.java
@@ -1,11 +1,10 @@
-package shop.annatation;
+package shop.validation;
 
 import jakarta.validation.Constraint;
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
-import shop.validator.FieldMatchValidator;
 
 @Constraint(validatedBy = FieldMatchValidator.class)
 @Target(ElementType.TYPE)
