+++ src/main/resources/db/changelog/changes/02-create-users-table.yaml
@@ -37,3 +37,4 @@ databaseChangeLog:
               - column:
                   name: shipping_address
                   type: varchar(255)
+
