+++ src/main/resources/db/changelog/db.changelog-master.yaml
@@ -1,4 +1,6 @@
 databaseChangeLog:
   - include:
       file: db/changelog/changes/01-create-books-table.yaml
+  - include:
+      file: db/changelog/changes/02-create-users-table.yaml
